refactor: 퇴장 로직을 ContextMenu에서 제거하고 CommunityLeaveBox로 분리
